+++ Scripts/New-FLCollectionsExplorer.ps1
@@ -1,20 +1,20 @@
 
-function New-FLDatabaseExplorer($ConnectionString, [switch]$System) {
+function New-FLCollectionsExplorer($ConnectionString, [switch]$System) {
 	New-Object PowerShellFar.PowerExplorer 2c77b1ed-c496-4823-9792-2e417a8c83a9 -Property @{
 		Data = @{
 			ConnectionString = $ConnectionString
 			System = $System
 		}
 		Functions = 'DeleteFiles, RenameFile'
-		AsCreatePanel = {FLDatabaseExplorer_AsCreatePanel @args}
-		AsDeleteFiles = {FLDatabaseExplorer_AsDeleteFiles @args}
-		AsExploreDirectory = {FLDatabaseExplorer_AsExploreDirectory @args}
-		AsGetFiles = {FLDatabaseExplorer_AsGetFiles @args}
-		AsRenameFile = {FLDatabaseExplorer_AsRenameFile @args}
+		AsCreatePanel = {FLCollectionsExplorer_AsCreatePanel @args}
+		AsDeleteFiles = {FLCollectionsExplorer_AsDeleteFiles @args}
+		AsExploreDirectory = {FLCollectionsExplorer_AsExploreDirectory @args}
+		AsGetFiles = {FLCollectionsExplorer_AsGetFiles @args}
+		AsRenameFile = {FLCollectionsExplorer_AsRenameFile @args}
 	}
 }
 
-function FLDatabaseExplorer_AsCreatePanel {
+function FLCollectionsExplorer_AsCreatePanel {
 	param($1)
 	$panel = [FarNet.Panel]$1
 	$panel.Title = 'Collections'
@@ -23,7 +23,7 @@ function FLDatabaseExplorer_AsCreatePanel {
 	$panel
 }
 
-function FLDatabaseExplorer_AsGetFiles($1) {
+function FLCollectionsExplorer_AsGetFiles($1) {
 	Use-LiteDatabase $1.Data.ConnectionString {
 		if ($1.Data.System) {
 			foreach($r in Invoke-LiteCommand 'SELECT $.name FROM $cols') {
@@ -38,11 +38,11 @@ function FLDatabaseExplorer_AsGetFiles($1) {
 	}
 }
 
-function FLDatabaseExplorer_AsExploreDirectory($1, $2) {
-	New-FLCollectionExplorer $1.Data.ConnectionString $2.File.Name
+function FLCollectionsExplorer_AsExploreDirectory($1, $2) {
+	New-FLDocumentsExplorer $1.Data.ConnectionString $2.File.Name
 }
 
-function FLDatabaseExplorer_AsRenameFile($1, $2) {
+function FLCollectionsExplorer_AsRenameFile($1, $2) {
 	$newName = ([string]$Far.Input('New name', $null, 'Rename', $2.File.Name)).Trim()
 	if (!$newName) {return}
 	Use-LiteDatabase $1.Data.ConnectionString {
@@ -51,7 +51,7 @@ function FLDatabaseExplorer_AsRenameFile($1, $2) {
 	$2.PostName = $newName
 }
 
-function FLDatabaseExplorer_AsDeleteFiles($1, $2) {
+function FLCollectionsExplorer_AsDeleteFiles($1, $2) {
 	# ask
 	if ($2.UI) {
 		$text = @"