+++ Scripts/New-FLDocumentsExplorer.ps1
@@ -1,32 +1,44 @@
 
-function New-FLCollectionExplorer($ConnectionString, $CollectionName) {
+function New-FLDocumentsExplorer($ConnectionString, $Query) {
+	$Query = $Query.Trim()
+	if ($Query -match '^\S+$') {
+		$CollectionName = $Query
+		$Query = $null
+	}
+	else {
+		$CollectionName = $null
+	}
 	New-Object PowerShellFar.ObjectExplorer -Property @{
 		Data = @{
 			ConnectionString = $ConnectionString
 			CollectionName = $CollectionName
+			Query = $Query
 		}
 		FileComparer = [PowerShellFar.FileMetaComparer]'_id'
-		AsCreateFile = {FLCollectionExplorer_AsCreateFile @args}
-		AsCreatePanel = {FLCollectionExplorer_AsCreatePanel @args}
-		AsDeleteFiles = {FLCollectionExplorer_AsDeleteFiles @args}
-		AsGetContent = {FLCollectionExplorer_AsGetContent @args}
-		AsGetData = {FLCollectionExplorer_AsGetData @args}
-		AsSetText = {FLCollectionExplorer_AsSetText @args}
+		AsCreateFile = {FLDocumentsExplorer_AsCreateFile @args}
+		AsCreatePanel = {FLDocumentsExplorer_AsCreatePanel @args}
+		AsDeleteFiles = {FLDocumentsExplorer_AsDeleteFiles @args}
+		AsGetContent = {FLDocumentsExplorer_AsGetContent @args}
+		AsGetData = {FLDocumentsExplorer_AsGetData @args}
+		AsSetText = {FLDocumentsExplorer_AsSetText @args}
 	}
 }
 
-function FLCollectionExplorer_AsCreatePanel($1) {
+function FLDocumentsExplorer_AsCreatePanel($1) {
 	$panel = [PowerShellFar.ObjectPanel]$1
-	$title = $1.Data.CollectionName
-
+	if ($1.Data.Query) {
+		$title = $1.Data.Query
+	}
+	else {
+		$title = $1.Data.CollectionName
+		$panel.PageLimit = 1000
+	}
 	$panel.Title = $title
-	$panel.PageLimit = 1000
-
 	$1.Data.Panel = $panel
 	$panel
 }
 
-function FLCollectionExplorer_EditorOpened {
+function FLDocumentsExplorer_EditorOpened {
 	$this.add_KeyDown({
 		if ($_.Key.KeyDown -and $_.Key.Is([FarNet.KeyCode]::F4)) {
 			$_.Ignore = $true
@@ -35,7 +47,7 @@ function FLCollectionExplorer_EditorOpened {
 	})
 }
 
-function FLCollectionExplorer_AsCreateFile($1, $2) {
+function FLDocumentsExplorer_AsCreateFile($1, $2) {
 	# edit new json
 	$json = ''
 	for() {
@@ -43,7 +55,7 @@ function FLCollectionExplorer_AsCreateFile($1, $2) {
 			Title = 'New document (JSON)'
 			Extension = 'js'
 			Text = $json
-			EditorOpened = {FLCollectionExplorer_EditorOpened}
+			EditorOpened = {FLDocumentsExplorer_EditorOpened}
 		}
 		$json = $Far.AnyEditor.EditText($arg)
 		if (!$json) {
@@ -78,12 +90,30 @@ function FLCollectionExplorer_AsCreateFile($1, $2) {
 	$1.Data.Panel.NeedsNewFiles = $true
 }
 
-function FLCollectionExplorer_AsDeleteFiles($1, $2) {
-	# ask
+function FLDocumentsExplorer_AsDeleteFiles($1, $2) {
+	# check _id
+	try {
+		foreach($doc in $2.FilesData) {
+			if ($null -eq $doc._id) {
+				throw 'Cannot edit documents without _id'
+			}
+		}
+	}
+	catch {
+		if ($2.UI) {
+			Show-FarMessage $_
+		}
+		return
+	}
+
+	# confirm
 	if ($2.UI) {
 		$text = "$($2.Files.Count) documents(s)"
-		if (Show-FarMessage $text Delete YesNo) {return}
+		if (Show-FarMessage $text Delete YesNo) {
+			return
+		}
 	}
+
 	# remove
 	try {
 		Use-LiteDatabase $1.Data.ConnectionString {
@@ -95,13 +125,22 @@ function FLCollectionExplorer_AsDeleteFiles($1, $2) {
 	}
 	catch {
 		$2.Result = 'Incomplete'
-		if ($2.UI) {Show-FarMessage "$_"}
+		if ($2.UI) {
+			Show-FarMessage "$_"
+		}
 	}
+
 	$1.Data.Panel.NeedsNewFiles = $true
 }
 
-function FLCollectionExplorer_AsGetContent($1, $2) {
+function FLDocumentsExplorer_AsGetContent($1, $2) {
 	$id = $2.File.Data._id
+	if ($null -eq $id) {
+		Show-FarMessage 'Cannot edit documents without _id'
+		$2.Result = 'Ignore'
+		return
+	}
+
 	$doc = Use-LiteDatabase $1.Data.ConnectionString {
 		$collection = Get-LiteCollection $1.Data.CollectionName
 		Get-LiteData $collection '$._id = @_id', @{_id = $id}
@@ -110,14 +149,21 @@ function FLCollectionExplorer_AsGetContent($1, $2) {
 	$2.UseText = $doc.Print()
 	$2.UseFileExtension = 'js'
 	$2.CanSet = $true
-	$2.EditorOpened = {FLCollectionExplorer_EditorOpened}
+	$2.EditorOpened = {FLDocumentsExplorer_EditorOpened}
 }
 
-function FLCollectionExplorer_AsGetData($1, $2) {
+function FLDocumentsExplorer_AsGetData($1, $2) {
 	if ($2.NewFiles -or !$1.Cache) {
 		Use-LiteDatabase $1.Data.ConnectionString {
-			$collection = Get-LiteCollection $1.Data.CollectionName
-			Get-LiteData $collection -First $2.Limit -Skip $2.Offset -As PS
+			if ($1.Data.Query) {
+				$name = ''
+				Invoke-LiteCommand $1.Data.Query -As PS -Collection ([ref]$name)
+				$1.Data.CollectionName = $name
+			}
+			else {
+				$collection = Get-LiteCollection $1.Data.CollectionName
+				Get-LiteData $collection -First $2.Limit -Skip $2.Offset -As PS
+			}
 		}
 	}
 	else {
@@ -125,7 +171,7 @@ function FLCollectionExplorer_AsGetData($1, $2) {
 	}
 }
 
-function FLCollectionExplorer_AsSetText($1, $2) {
+function FLDocumentsExplorer_AsSetText($1, $2) {
 	$new = [Ldbc.Dictionary]::FromJson($2.Text)
 	$new.EnsureId()