+++ Release-Notes.md
@@ -1,5 +1,14 @@
 # FarLite Release Notes
 
+## v0.0.2
+
+Requires Ldbc 0.4.2
+
+`Open-LitePanel` supports queries in addition to collection names using the
+same parameter `Query`. If queried documents contain `_id` keys then documents
+are edited and deleted as usual. Paging is not supported for query results but
+queries may use `LIMIT` and `OFFSET` themselves.
+
 ## v0.0.1
 
 Requires Ldbc 0.4.1